The load balancer's VIP never answered
Book §23.6 has students curl a virtual IP that gini.samples.l4_lb spreads across two
web servers. It failed every time, for four independent reasons, each of which alone
was enough and each of which hid the next.

1. Nothing answered ARP for the VIP. No host owns 10.0.1.100, so the client's
   "who has 10.0.1.100?" went unanswered and it never sent an IP packet; the flow
   rules never ran. The controller now replies with a made-up MAC (--vmac), which is
   what "advertises a VIP" means in practice.

2. It forwarded through OFPP_NORMAL. On the GINI Flow Switch (the gRouter in
   --openflow mode) NORMAL means "hand to the router's own IP pipeline", not L2
   switching, and on a flat segment the router has no interfaces or routes to use.
   With l4_lb loaded even a DIRECT fetch of a backend failed, because ARP between
   hosts never crossed the switch. Non-VIP traffic is now forwarded the way
   gini.samples.switch does it: learn MAC -> port, flood the unknown. (A packet-out
   for a frame that came from a packet-in must not name in_port=OFPP_NONE; that
   overrode the ingress port POX copies in, and broke ARP outright.)

3. Only the IP was rewritten, so the frame stayed addressed to the VIP's MAC and no
   backend's NIC accepted it. The forward flow now sets dl_dst+nw_dst, the reverse
   flow dl_src+nw_src, and both sit above a learned MAC rule's priority.

4. Underneath, the switch crashed on the first TCP address rewrite: SIGSEGV in
   checksum() from tcp_checksum(). tcp_checksum/udp_checksum stepped a typed pointer
   (ip_packet + hdr_len*4 moves hdr_len*4 whole structs), summed &tcp_packet (the
   bytes of the pointer variable on the stack), and mixed network- and host-order
   lengths. And tcp.h declared seq_num/ack_num as uint16_t, putting `checksum` at
   byte 12 instead of 16, so even a correct sum would have been written over
   data-offset/flags/window. This hit EVERY OpenFlow SET_NW_SRC/DST of a TCP packet,
   not just the load balancer. Both functions are now thin wrappers over one
   endian-safe l4_checksum() in utils.c that bounds-checks the length and keeps
   UDP's "0 means none" / "computed 0 goes out as 0xFFFF" rules.

Verified against real containers on the Figure 23.2 topology, handout steps 4-8:
Client_1 lands on Web_1 and stays, Client_2 on Web_2; `openflow entry all` shows
the two forward and two reverse rules with their rewrite actions; the counters move
only for the client fetching; the controller log shows the round-robin assignment;
stopping Web_1's server gives Client_1 an immediate "connection refused" while
Client_2 still answers (the no-health-checking lesson).

With GINI's automatic addressing, hosts are numbered .10, .11, ... in placement
order, so the app's default backends .11,.12 are Client_2 and Web_1. The controller
now logs a warning when a pool member sends to the VIP, so a student sees why one
client fails instead of nothing at all. The handout's step 1 should say so too.

checksum_test.c pins the gRouter checksums against values from an independent
RFC 1071 implementation in Python (even/odd TCP, the SET_NW_DST rewrite itself, UDP,
UDP "none", a bogus length); test_l4_lb.py pins the app's shape. Both images need
rebuilding for students: gini-grouter (C) and gini-pox (the app is baked in).

+++ backend/include/tcp.h
@@ -48,9 +48,12 @@ typedef struct {
 	uint16_t	src_port;
 	// Destination port
 	uint16_t	dst_port;
-	// Sequence number
-	uint16_t	seq_num;
-	uint16_t	ack_num;
+	// Sequence and acknowledgement numbers — 32 bits each on the wire. They were
+	// declared uint16_t, which put `checksum` at byte 12 instead of 16: a rewritten TCP
+	// packet then had its checksum written over data-offset/flags/window, mangling the
+	// header of every flow an OpenFlow SET_NW_* rule touched.
+	uint32_t	seq_num;
+	uint32_t	ack_num;
 #if __BYTE_ORDER == __LITTLE_ENDIAN
 	// Unused
 	uint8_t		reserved:4;
